+++ ShareCad.Core/Images/connect_icon.png
