+++ ShareCad.Core/Properties/Resources.resx
