+++ ShareCad.Core/ShareCad.Core.csproj
@@ -1,10 +1,10 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>ed1b550b-94b6-45c0-99d6-57715a8de7bb</ProjectGuid>
+    <ProjectGuid>{ED1B550B-94B6-45C0-99D6-57715A8DE7BB}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>ShareCad.Core</RootNamespace>
@@ -31,24 +31,96 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="System"/>
-    
-    <Reference Include="System.Core"/>
-    <Reference Include="System.Xml.Linq"/>
-    <Reference Include="System.Data.DataSetExtensions"/>
-    
-    
-    <Reference Include="Microsoft.CSharp"/>
-    
-    <Reference Include="System.Data"/>
-    
-    <Reference Include="System.Net.Http"/>
-    
-    <Reference Include="System.Xml"/>
+    <Reference Include="0Harmony">
+      <HintPath>..\Libs\0Harmony.dll</HintPath>
+    </Reference>
+    <Reference Include="PresentationCore" />
+    <Reference Include="PresentationFramework" />
+    <Reference Include="Ptc.Calculation">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Calculation.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Calculation.Core">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Calculation.Core.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls.CollapsibleArea">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.CollapsibleArea.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls.Core">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Core.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls.EquationControl">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.EquationControl.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls.ExcelComponent">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.ExcelComponent.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls.Include">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Include.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls.Plots">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Plots.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls.SolveBlock">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.SolveBlock.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls.Text">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Text.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls.Whiteboard">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Whiteboard.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Controls.Worksheet">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Worksheet.dll</HintPath>
+    </Reference>
+    <Reference Include="Ptc.Core">
+      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Core.dll</HintPath>
+    </Reference>
+    <Reference Include="System" />
+    <Reference Include="System.Core" />
+    <Reference Include="System.Xml.Linq" />
+    <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="Microsoft.CSharp" />
+    <Reference Include="System.Data" />
+    <Reference Include="System.Net.Http" />
+    <Reference Include="System.Xml" />
+    <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Class1.cs" />
+    <Compile Include="BootlegResourceManager.cs" />
+    <Compile Include="Core.cs" />
+    <Compile Include="ManipulateWorksheet.cs" />
+    <Compile Include="MathcadHelper\CustomMcdxDeserializer.cs" />
+    <Compile Include="MathcadHelper\CustomMcdxSerializer.cs" />
+    <Compile Include="MathcadHelper\CustomWorksheetSectionDeserializationStrategy.cs" />
+    <Compile Include="MathcadHelper\CustomWorksheetSectionSerializationStrategy.cs" />
+    <Compile Include="MathcadHelper\WorksheetControlExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Properties\Resources.Designer.cs">
+      <DependentUpon>Resources.resx</DependentUpon>
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+    </Compile>
+    <Compile Include="SharedDocument.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Images\connect_icon.png">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Images\noconnect_icon.png">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Properties\Resources.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
+      <SubType>Designer</SubType>
+    </EmbeddedResource>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
- </Project>
+</Project>
