+++ ShareCad/ShareCad.csproj
@@ -80,48 +80,6 @@
     </Reference>
     <Reference Include="PresentationCore" />
     <Reference Include="PresentationFramework" />
-    <Reference Include="Ptc.Calculation">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Calculation.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Calculation.Core">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Calculation.Core.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls.CollapsibleArea">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.CollapsibleArea.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls.Core">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Core.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls.EquationControl">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.EquationControl.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls.ExcelComponent">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.ExcelComponent.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls.Include">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Include.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls.Plots">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Plots.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls.SolveBlock">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.SolveBlock.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls.Text">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Text.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls.Whiteboard">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Whiteboard.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Controls.Worksheet">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Controls.Worksheet.dll</HintPath>
-    </Reference>
-    <Reference Include="Ptc.Core">
-      <HintPath>..\..\..\..\..\..\Program Files\PTC\Mathcad Prime 6.0.0.0\Ptc.Core.dll</HintPath>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.ComponentModel.DataAnnotations" />
     <Reference Include="System.Core" />
@@ -142,28 +100,19 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BootlegResourceManager.cs" />
     <Compile Include="Old stuff\ManipulateSingleRegionsOld.cs" />
-    <Compile Include="MathcadHelper\CustomMcdxDeserializer.cs" />
-    <Compile Include="ManipulateWorksheet.cs" />
-    <Compile Include="MathcadHelper\CustomMcdxSerializer.cs" />
-    <Compile Include="MathcadHelper\CustomWorksheetSectionSerializationStrategy.cs" />
-    <Compile Include="Properties\Resources.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-      <DependentUpon>Resources.resx</DependentUpon>
-    </Compile>
     <Compile Include="ShareCad.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SharedDocument.cs" />
     <Compile Include="WinConsole.cs" />
-    <Compile Include="MathcadHelper\WorksheetControlExtensions.cs" />
-    <Compile Include="MathcadHelper\CustomWorksheetSectionDeserializationStrategy.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\ShareCad.Core\ShareCad.Core.csproj">
+      <Project>{ed1b550b-94b6-45c0-99d6-57715a8de7bb}</Project>
+      <Name>ShareCad.Core</Name>
+    </ProjectReference>
     <ProjectReference Include="..\ShareCad.Logging\ShareCad.Logging.csproj">
       <Project>{6D792E28-01A9-401B-B57B-5347BF9A3E34}</Project>
       <Name>ShareCad.Logging</Name>
@@ -176,23 +125,6 @@
       <Project>{bcdaa12c-45ac-45cd-8961-9d11b6917153}</Project>
       <Name>ShareCad.UI</Name>
     </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Properties\Resources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <SubType>Designer</SubType>
-      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Images\connect_icon.png">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="Images\noconnect_icon.png">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
   </ItemGroup>
   <ItemGroup>
     <Analyzer Include="..\packages\StyleCop.Analyzers.1.1.118\analyzers\dotnet\cs\StyleCop.Analyzers.CodeFixes.dll" />
